requirements.txt
sqlalchemy
flask
strawberry-graphql
jinja2